mobiDB_human.jsonのkey名を修正。 拡張子を修正。mjson->json
