+++ sample_program/json_related/json_rename_key.py
@@ -1,13 +1,3 @@
-from bioservices import UniProt  # Uniprotのメソッドをインポート
-import MySQLdb
-import logging
-import threading
-import time
-import itertools
-
-from bioservices import UniProt  # Uniprotのメソッドをインポート
-import pandas as pd
-import io
 from logging import getLogger, StreamHandler, DEBUG
 import time
 import json
@@ -24,18 +14,12 @@ logger.propagate = False
 class InfoAddName:
     def search_info(self):
         ar = []
-        with open('disorder_add_protain.mjson', 'a') as fw:
-            with open('disorder.mjson', 'r') as fr:
-                # with open('disorder.mjson', 'r') as fr:
+        with open('mobiDB_human.json', 'a') as fw:
+            with open('mobiDB_human.mjson', 'r') as fr:
                 for (i, line) in enumerate(fr):
-                    print(i)
                     try:
                         json_dict = json.loads(line)
-                        result = service.search(json_dict["acc"], frmt='tab', columns=columnlist)
-                        result_s = result.splitlines()
-
-                        a_dict = {"protein names": result_s[1]}
-                        json_dict.update(a_dict)
+                        # json_dict['protein_names'] = json_dict.pop("protein names")
 
                         fw.write('{}\n'.format(json.dumps(json_dict)))
 
@@ -50,21 +34,17 @@ class InfoAddName:
 
                 print(ar)
 
-        logging.debug('worker End')
+        logger.debug('worker End')
 
 
 if __name__ == '__main__':
     # データをDBに格納する
-    with open('disorder_add_protain.mjson', 'w'):
+    with open('mobiDB_human.json', 'w'):
         pass
 
-    logging.debug('main Begin')
+    logger.debug('main Begin')
 
     t1 = time.time()
-
-    service = UniProt()
-    columnlist = 'protein names'
-    # columnlist = 'keywords'
 
     idn = InfoAddName()
     idn.search_info()
@@ -72,7 +52,5 @@ if __name__ == '__main__':
     t2 = time.time()
     elapsed_time = t2 - t1  # 処理にかかった時間を計算する
     print("経過時間：", elapsed_time)
-
-    logging.debug('main End')
